replace setposixattributes.py with bulkload version
